refator: Toast support gesture.
