+++ packages/design-system/toast/index.tsx
@@ -1,17 +1,7 @@
 import { useState, useRef, createContext, useMemo, useContext } from "react";
-import {
-  Platform,
-  View,
-  AccessibilityInfo,
-  LayoutChangeEvent,
-  StyleSheet,
-} from "react-native";
+import { AccessibilityInfo } from "react-native";
 
-import { MotiView, AnimatePresence } from "moti";
-
-import { useSafeAreaInsets } from "@showtime-xyz/universal.safe-area";
-import { tw } from "@showtime-xyz/universal.tailwind";
-import { Text } from "@showtime-xyz/universal.text";
+import { ToastContainer } from "./toast-container";
 
 type ShowParams = {
   message?: string;
@@ -19,28 +9,19 @@ type ShowParams = {
   element?: React.ReactElement;
 };
 
-type ToastContext = {
+type ToastContextType = {
   show: (params: ShowParams) => void;
   hide: () => void;
   isVisible: boolean;
 };
 
-// eslint-disable-next-line no-redeclare
-const ToastContext = createContext<ToastContext | undefined>(undefined);
-
-const SAFE_AREA_TOP = 20;
+const ToastContext = createContext<ToastContextType | undefined>(undefined);
 
 export const ToastProvider = ({ children }: any) => {
   const [show, setShow] = useState(false);
   const [message, setMessage] = useState<string | undefined>();
-  const [render, setRender] = useState<React.ReactElement | null>();
-  const [layout, setLayout] = useState<
-    LayoutChangeEvent["nativeEvent"]["layout"] | undefined
-  >();
+  const [render, setRender] = useState<JSX.Element | null>();
   const hideTimeoutRef = useRef<any>(null);
-  const safeAreaInsets = useSafeAreaInsets();
-  const { top: safeAreaTop } =
-    Platform.OS === "web" ? { top: SAFE_AREA_TOP } : safeAreaInsets;
 
   const value = useMemo(
     () => ({
@@ -72,71 +53,18 @@ export const ToastProvider = ({ children }: any) => {
     [show, setShow]
   );
 
-  const toastHeight = layout?.height ?? 0;
-
   return (
     <ToastContext.Provider value={value}>
       {children}
-      <AnimatePresence>
-        {show ? (
-          <View
-            style={[
-              StyleSheet.absoluteFill,
-              //@ts-ignore
-              // TODO: improve toast to attach to react portal!
-              Platform.OS === "web" ? { position: "fixed", zIndex: 999 } : {},
-            ]}
-            pointerEvents="box-none"
-          >
-            <MotiView
-              style={[
-                styles.toastContainer,
-                { opacity: layout ? 1 : 0 },
-                tw.style(
-                  "bg-white dark:bg-black shadow-black dark:shadow-white"
-                ),
-              ]}
-              accessibilityLiveRegion="polite"
-              pointerEvents="box-none"
-              from={{ translateY: -toastHeight }}
-              animate={{
-                translateY: safeAreaTop === 0 ? SAFE_AREA_TOP : safeAreaTop,
-              }}
-              exit={{ translateY: -toastHeight }}
-              transition={{ type: "timing", duration: 350 }}
-              onLayout={(e) => {
-                setLayout(e.nativeEvent.layout);
-              }}
-            >
-              {render ? (
-                render
-              ) : (
-                <Text tw="p-4 text-center text-gray-900 dark:text-white">
-                  {message}
-                </Text>
-              )}
-            </MotiView>
-          </View>
-        ) : null}
-      </AnimatePresence>
+      <ToastContainer
+        show={show}
+        hide={value.hide}
+        message={message}
+        render={render}
+      />
     </ToastContext.Provider>
   );
 };
-
-const styles = StyleSheet.create({
-  toastContainer: {
-    alignSelf: "center",
-    borderRadius: 16,
-    shadowOffset: {
-      width: 0,
-      height: 8,
-    },
-    shadowOpacity: 0.1,
-    shadowRadius: 16,
-    elevation: 5,
-    justifyContent: "center",
-  },
-});
 
 export const useToast = () => {
   const toast = useContext(ToastContext);
