+++ packages/design-system/toast/toast.tsx
@@ -1,0 +1,75 @@
+import { useState } from "react";
+import { Platform, LayoutChangeEvent, StyleSheet } from "react-native";
+
+import { MotiView } from "moti";
+
+import { useSafeAreaInsets } from "@showtime-xyz/universal.safe-area";
+import { tw } from "@showtime-xyz/universal.tailwind";
+import { Text } from "@showtime-xyz/universal.text";
+
+import { PanToClose } from "../pan-to-close";
+
+type ToastProps = {
+  render?: JSX.Element | null;
+  message?: string;
+  hide: () => void;
+};
+
+const SAFE_AREA_TOP = 20;
+
+export const Toast = ({ render, message, hide }: ToastProps) => {
+  const safeAreaInsets = useSafeAreaInsets();
+  const [layout, setLayout] = useState<
+    LayoutChangeEvent["nativeEvent"]["layout"] | undefined
+  >();
+  const { top: safeAreaTop } =
+    Platform.OS === "web" ? { top: SAFE_AREA_TOP } : safeAreaInsets;
+
+  const toastHeight = layout?.height ?? 0;
+
+  return (
+    <PanToClose panCloseDirection={"up"} onClose={hide}>
+      <MotiView
+        style={[
+          styles.toastContainer,
+          { opacity: layout ? 1 : 0 },
+          tw.style("bg-white dark:bg-black shadow-black dark:shadow-white"),
+        ]}
+        accessibilityLiveRegion="polite"
+        pointerEvents="box-none"
+        from={{ translateY: -toastHeight }}
+        animate={{
+          translateY: safeAreaTop === 0 ? SAFE_AREA_TOP : safeAreaTop,
+        }}
+        exit={{ translateY: -toastHeight }}
+        transition={{ type: "timing", duration: 350 }}
+        onLayout={(e) => {
+          setLayout(e.nativeEvent.layout);
+        }}
+      >
+        {render ? (
+          render
+        ) : (
+          <Text tw="p-4 text-center text-gray-900 dark:text-white">
+            {message}
+          </Text>
+        )}
+      </MotiView>
+    </PanToClose>
+  );
+};
+
+const styles = StyleSheet.create({
+  toastContainer: {
+    alignSelf: "center",
+    borderRadius: 16,
+    shadowOffset: {
+      width: 0,
+      height: 8,
+    },
+    shadowOpacity: 0.1,
+    shadowRadius: 16,
+    elevation: 5,
+    justifyContent: "center",
+  },
+});
